+++ src/forms/upload/upload-button/upload-button.component.js
@@ -7,6 +7,9 @@ const UploadButton = {
   bindings: {
     name: '<',
     label: '<',
+    isSecondary: '<',
+    small: '<',
+    block: '<',
     accept: '<',
     model: '<',
     multiple: '<',