+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v5.1.0
+## Multi upload component
+Add multi upload component
+
 # v5.0.2
 ## Minor problem fixes for upload component
 Added async apply to some UI state changes in live upload component